fix: setting multiple cookies in streaming mode makes an invalid `set-cookie` header.
For example, when using Astro Auth multiple headers can be set causing them to be concatenated with a comma. This is not a valid `Set-Cookie` header so only the first cookie is honored and the others are discarded.

This causes an issue where certain auth cookies are not set eventually causing errors preventing users from logging in.